refactor odds
