+++ sportsbooklib/calculators/parlay_calc.py
@@ -3,7 +3,7 @@ from math import prod
 from typing import List, Mapping, Union
 from sportsbooklib.models.exceptions import InvalidSelectionsInputException
 from sportsbooklib.models.odds.enums import OddsFormat
-from sportsbooklib.models.odds.odds import Odds
+from sportsbooklib.models.odds import Odds
 
 from sportsbooklib.models.selection.selection import Selection
