+++ tests/calculators/test_parlay_calc.py
@@ -5,7 +5,7 @@ from decimal import Decimal
 
 from sportsbooklib.models.exceptions import InvalidSelectionsInputException
 from sportsbooklib.models.odds.enums import OddsFormat
-from sportsbooklib.models.odds.odds import Odds
+from sportsbooklib.models.odds import Odds
 from sportsbooklib.models.selection.selection import Selection
